Feature/healthpick

+++ ddjd-fp/Assets/Scripts/Player/StateMachine/Ability/SubStates/PlayerInteractState.cs
@@ -22,12 +22,14 @@ public class PlayerInteractState : PlayerAbilityState {
 
     private void HandleObject() {
         if(_context.InteractableItem != null) {
-            if (_context.InteractableItem.tag == "Health") {
-                Events.OnCatchHealthPlant.Invoke();
+            if (_context.InteractableItem.tag == "Health" && _context.Data.HealthCrystal < _context.Data.MaxHealthCrystal) {
+                Events.OnCatchHealthCrystal.Invoke();
                 _context.DestroyObject(_context.InteractableItem);
+                _context.GetItem(0);
             } else if (_context.InteractableItem.tag == "Crystal") {
-                Events.OnCatchCrystal.Invoke();
+                Events.OnCatchManaCrystal.Invoke();
                 _context.DestroyObject(_context.InteractableItem);
+                _context.GetItem(1);
             }    
         }
